+++ src/shared/infrastructure/http/app.ts
@@ -326,6 +326,8 @@ export const createExpressApp = (
       controllers.getMyCollectionsController,
       controllers.getCollectionsController,
       controllers.getCollectionsForUrlController,
+      controllers.getRecommendedCollectionsForUrlController,
+      controllers.getRecommendedOpenCollectionsForUrlController,
       controllers.searchCollectionsController,
       controllers.getOpenCollectionsWithContributorController,
       controllers.getCollectionFollowersController,