+++ src/shared/infrastructure/http/factories/UseCaseFactory.ts
@@ -40,6 +40,8 @@ import { GetCollectionPageByAtUriUseCase } from 'src/modules/cards/application/u
 import { GetUrlStatusForMyLibraryUseCase } from '../../../../modules/cards/application/useCases/queries/GetUrlStatusForMyLibraryUseCase';
 import { GetLibrariesForUrlUseCase } from '../../../../modules/cards/application/useCases/queries/GetLibrariesForUrlUseCase';
 import { GetCollectionsForUrlUseCase } from '../../../../modules/cards/application/useCases/queries/GetCollectionsForUrlUseCase';
+import { GetRecommendedCollectionsForUrlUseCase } from '../../../../modules/cards/application/useCases/queries/GetRecommendedCollectionsForUrlUseCase';
+import { GetRecommendedOpenCollectionsForUrlUseCase } from '../../../../modules/cards/application/useCases/queries/GetRecommendedOpenCollectionsForUrlUseCase';
 import { GetNoteCardsForUrlUseCase } from '../../../../modules/cards/application/useCases/queries/GetNoteCardsForUrlUseCase';
 import { GetConnectionsForUrlUseCase } from '../../../../modules/cards/application/useCases/queries/GetConnectionsForUrlUseCase';
 import { GetConnectionsUseCase } from '../../../../modules/cards/application/useCases/queries/GetConnectionsUseCase';
@@ -176,6 +178,8 @@ export interface UseCases {
   getUrlStatusForMyLibraryUseCase: GetUrlStatusForMyLibraryUseCase;
   getLibrariesForUrlUseCase: GetLibrariesForUrlUseCase;
   getCollectionsForUrlUseCase: GetCollectionsForUrlUseCase;
+  getRecommendedCollectionsForUrlUseCase: GetRecommendedCollectionsForUrlUseCase;
+  getRecommendedOpenCollectionsForUrlUseCase: GetRecommendedOpenCollectionsForUrlUseCase;
   getNoteCardsForUrlUseCase: GetNoteCardsForUrlUseCase;
   getConnectionsForUrlUseCase: GetConnectionsForUrlUseCase;
   // Connection use cases
@@ -510,6 +514,18 @@ export class UseCaseFactory {
         repositories.collectionRepository,
         repositories.followsRepository,
       ),
+      getRecommendedCollectionsForUrlUseCase:
+        new GetRecommendedCollectionsForUrlUseCase(
+          services.searchService,
+          repositories.collectionQueryRepository,
+          services.profileService,
+        ),
+      getRecommendedOpenCollectionsForUrlUseCase:
+        new GetRecommendedOpenCollectionsForUrlUseCase(
+          services.searchService,
+          repositories.collectionQueryRepository,
+          services.profileService,
+        ),
       getNoteCardsForUrlUseCase: new GetNoteCardsForUrlUseCase(
         repositories.cardQueryRepository,
         services.profileService,